backend/godapp_main/pyproject.toml
[project]
name = "godapp_main"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleksandr",email = "jirafster@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=6.0,<7.0)",
    "django-ninja (>=1.5.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "django-polymorphic (>=4.4.1,<5.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
